fix(orm): replace Double with Numeric for monetary and quantity columns
purchases.total, purchase_details.unit_price → Numeric(12,2)
purchase_details.quantity → Numeric(10,4)
products.price_retail, price_cost → Numeric(12,2)
Prevents floating-point rounding errors in financial calculations.

+++ src/pharmatrack/models/purchase_details/orm.py
@@ -1,8 +1,9 @@
-from sqlalchemy import BigInteger, Double, Date, TIMESTAMP, String, ForeignKey
+from sqlalchemy import BigInteger, Numeric, Date, TIMESTAMP, String, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import func
 from ...db.session import Base
 from datetime import datetime, date
+from decimal import Decimal
 from typing import Optional
 
 
@@ -22,8 +23,8 @@ class PurchaseDetail(Base):
         nullable=False,
     )
 
-    quantity: Mapped[float] = mapped_column(Double, nullable=False)
-    unit_price: Mapped[float] = mapped_column(Double, nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
 
     expiration_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
     lot_code: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
